[NUGUDI-168] refactor(web): logger를 shared/utils로 이동
- src/lib/logger.ts를 src/shared/utils/logger.ts로 이동
- 전역 유틸리티는 shared 레이어에 위치하도록 구조 개선
- 모든 import 경로 업데이트
