+++ apps/web/src/shared/configs/pwa/register-service-worker.ts
@@ -1,4 +1,4 @@
-import { logger } from "@/src/lib/logger";
+import { logger } from "@/src/shared/utils/logger";
 
 export const registerServiceWorker = () => {
   if (typeof window === "undefined") return;
